server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.12.14",
    "aiosignal==1.4.0",
    "alembic==1.16.4",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "async-timeout==5.0.1",
    "attrs==25.3.0",
    "bcrypt==4.3.0",
    "cffi==1.17.1",
    "click==8.1.8",
    "cryptography==45.0.5",
    "ecdsa==0.19.1",
    "exceptiongroup==1.3.0",
    "fastapi==0.116.1",
    "frozenlist==1.7.0",
    "h11==0.16.0",
    "httptools==0.6.4",
    "idna==3.10",
    "mako==1.3.10",
    "markupsafe==3.0.2",
    "multidict==6.6.3",
    "numpy==2.0.2",
    "passlib==1.7.4",
    "propcache==0.3.2",
    "pyasn1==0.6.1",
    "pycparser==2.22",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pydantic-settings==2.10.1",
    "python-dotenv==1.1.1",
    "python-jose==3.5.0",
    "python-multipart==0.0.20",
    "pyyaml==6.0.2",
    "rsa==4.9.1",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.41",
    "starlette==0.47.2",
    "tomli==2.2.1",
    "typing-extensions==4.14.1",
    "typing-inspection==0.4.1",
    "uvicorn==0.35.0",
    "uvloop==0.21.0",
    "watchfiles==1.1.0",
    "websockets==15.0.1",
    "yarl==1.20.1",
]
